configfile: move related logic into ConfigFile

+++ tests/conftest.py
@@ -10,13 +10,14 @@ def set_pcw_ini(filename, add=''):
         f.write(add)
 
 
-@pytest.fixture(scope="session", autouse=True)
+@pytest.fixture(autouse=True)
 def pcw_file():
     tmpFile = tempfile.mkstemp()
     webui.settings.CONFIG_FILE = tmpFile[1]
     set_pcw_ini(tmpFile[1])
     yield tmpFile[1]
-    os.remove(tmpFile[1])
+    if os.path.exists(tmpFile[1]):
+        os.remove(tmpFile[1])
